requirements.txt
fastapi
uvicorn[standard]
paho-mqtt
SQLAlchemy
mysql-connector-python
python-dotenv
pydantic
